Quite a big syntaxic patch, do not pay attention to the number of lines ! Nothing new with this commit, but small fixes.
Fixed:
	- Big parenthesis did not renderer weirdly with "overlaping" looking parts that was due to
	  antialiasing.

Changed:
	- Added a lot of comments in the code, and removed all underscores `_` in method names.
	- all .h have been renamed into .hpp
	- a header/src now has the same name that the class it codes for.
	- the screen size has been set to 320 x 240: the real one.

I did not respect some convention on fracrenderer., calculator. and button., because
this is just code for the demo.  The only code that will be exported for the future is
the EditionTree. family. I'm currently working to make it a static library that could be
imported in another Qt project :)