Calculation sheet - like interface.
To say that it is a working calc. sheet would be too much. For now, on 'EXE',
the current edited edition tree send a std::string corresponding to his
expression to a parser, that check its syntax and evaluate it. It is not yet
evaluated using SAGE, and does not support square roots. Furthermore, errors
messages are not self-explanatory.

Fixed/Modified:

	- 'MathRenderer' renamed after 'CalcSheet' and many others renamed,
	- all graphical parameters have been regrouped into one folder
'editiontree/graphicalParameters.hpp' that can be modified at before compilation.
	- improved visuals on fractions (fraction bar more balanced, its size is
more appropriate when empty)
	- improved visuals on square roots, it is now more scalable.

Added:
	- 'CalcSheet' own a list of expressions and results that contains an
historic of the previous computations. This list can be arbitrarly big.
	- the edited area is separated from the rendering of the historic. The
cursor is for now stuck into it and the user cannot import expressions from
previous computations yet.