Communicating with Sage! This is a first version, in which there are two processes communicating through piped stdin-stdout.
One is the calculation sheet, and the other is a home-made Sage
application. On 'EXE', the expression tree gives a string corresponding to
its expression to the parser, that checks the syntax. If the syntax is accepted,
the expression is sent through stdin to the Sage appli, that calls 'sage_eval'
to evaluate the expression, and returns through stdout a raw string as a result.