integrate popups and a function bar into SLWindow
This change introduces a function bar used to display (so far, not
handle) functions for the F1..F5 keys. The function bar has a model
that is changed dynamically when changing screens, and can be overriden
by any component in the focus chain.

Currently the overlay still covers the function bar.

Co-authored-by: Nicolas Chappe <nicolas.chappe@ens-lyon.fr>