use structured objects as function bar models
This change loses the array-like property of the model because QML
components are now used instead of variants. A property of type
list<SLFunctionKeyModel> can be created, but its elements cannot be
assigned independently, as in

  SLFunctionBarModel {
    f[5]: SLFunctionKeyModel {
      text: "Back"
    }
  }

So properties named f1..f5 are used and the Repeater in SLFunctionBar
has been dropped for now.