update overlay height based on itself instead of the window size
Apparently there are situations in which the overlay's updateGeometry()
is called but the binding is not re-evaluated. Monitor the overlay's
height itself to make sure no change is missed.

Also, because we can't bind overlay.height based on itself, it needs to
be changed through JS so the binding to overlay.parent is lost. This is
handled with a second signal.