AppletHelper: update model when new window is set
When applets are started, the new active focus is determined before the
window is set in the newly-created AppletHelper, so updateModel() is not
called until focus changes after the application is opened.

This happens immediately in Funapp because of the SLStackLayout, which
is why this went unnoticed. Calcsheet has a simpler model where active
focus does not change until user action after opening the applet. As a
result, the function bar of the ExprField would be invisible.

This change forces a model update when the window changes, which
includes a check at startup, to solve this issue.