edition trees: basic support for sum operations
The term to sum is added between parentheses as a child of the sum node
so there is no need to capture it in the flow.

Giac wants the bound variable name to be separated from the lower bound
so currently a quick-and-dirty parsing of "x:=value" is used.

Rendering-wise, the size of the sigma symbol cannot be obtained easily
with the bounding rectangle so it is hardcoded so that everything ties
together properly. There are stil hacks lying because the size of the
text is hard to estimate as usual.