funapp: synchronize window range popup with graph's view
The update of the graph's view rectangle previously relied on
activeFocus and visible changes to pull the coordinates specified in the
window range popup. This resulted in the popup's input being overridden
by the current view when the graph tab is selected.

This commit changes the update mechanism to change the graph's view when
the window range popup is closed.