ide: fix runtime warnings in  file list when executing programs
File list items trigger a warning because of the [height: parent.height]
binding failing when executing files in the IDE, with parent = null.

It is likely that the write to a temporary file is causing the file list
model to reload, and the items to be removed from their parents before
being disposed of.