ide: use a pause script to avoid closing terminal early
Kind of hackish in a way. The script basically runs the interpreter
command then some [read -p] to stall the terminal.

One idea was to catch finished() and then bind termKeyPressed() to
close() to implement the delay. But after finished() the terminal is
blank. setAutoClose(false) disables this behavior but also blocks the
finish() signal.