ide: add the save functionality
This change replaces the "Cycle font size" key (which can come back
later) with the "Save" key, that changes its message to "Save*" when
there are unsaved changes and saves when pressed.

This is implemented with the saveAs() method of SourceEditor and the
modified attribute of QTextDocument, which is manually exposed to QML
because QQuickTextDocument hides everything from QML (as per the
documentation: "The class has to be used from C++ directly").

There are currently no safeguards to prevent executing an unsaved file
or leaving the editor with unsaved modifications.