graph, funapp: catch and log parsing and evaluation errors
Error messages are sent to the console, and corresponding data points
are not shown.

There is a very strange occurrence where a function with runtime errors,
such as 1/"A", fills up the sst_at_stack vector used for single-step
execution of Giac programs, causing errors to occur for other functions.