funapp: show errors in a popup after plotting
This changed adds an error report mechanism for the graph view. The
SLQuickGraph class is extended with:

* An mErrors member storing the list of current errors to be reported to
  the user.
* Two methods addError() and clearErrors() used when syntax or
  evaluation errors are reported by giac, and before a replot,
  respectively.

The replot mechanism has been changed from the onVisible signal of the
graph view to the explicit moment when the user switches to it. This is
because the order of onVisible being activated and the active focus
being forced to the graph as part of the tab changing mechanism in
SLStackLayout is not easy to guarantee. In particular, if the error
popup is shown automatically, then the forceActiveFocus() call takes
place *after* the popup is open and unpleasantly moves the focus away.

Because all functions are shown by default, Edition Trees that produce
empty Giac code are filtered out of the plot without an error.

Finally, an errorPopup has been added to the application. When error
messages have been recorded, the F1 key in the graph view reads
"Errors!" and is used to open the errorPopup which shows the details.