ide: improve key navigation in new file menu
KeyNavigation.priority also applies when the combo box is open, meaning
that the combo box can no longer be scrolled through when it is
beforeItem.

This change works around that leaving the priority to its default and
filtering Down presses when the combo box is closed to move focus
manually (KeyNavigation.down is still set because it is used when moving
up from the file name).