ide: separate KSyntaxHighlighting and languages.json names
This change adds a khighlight attribute to languages in languages.json
that specifies the definition to use for syntax highlighting. This
attribute is required for languages specified in languages.json to have
any highlighting.

In the event that a file is loaded that does not match any extension in
languages.json, no language data is loaded in the IDE (most notably, no
catalog will be available) but syntax highlighting is still enabled if
detected from the file name by KSyntaxHighlighting.