lib/graph: refuse to set invalid window ranges
This would cause the m_view to desync from the QCustomPlot axis ranges,
causing confusion as to what the true limits are.

Also this change means that the range is (safe for other omissions in
the code) always valid, hence loading/leaving the application should
always output a valid functions.json file.